feat: rename WorkspaceApiClientBase to WorkspaceApiClient and WorkspaceApiClient to WorkspaceApiClientImplement
